art-rl/requirements.txt
# Core dependencies
zenml>=0.70.0

# ART (Agentic Reinforcement Training) with LangGraph integration
openpipe-art[backend,langgraph]>=0.4.11

# LangChain/LangGraph for agent framework
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0

# LLM inference
litellm>=1.50.0

# Data handling
datasets>=3.0.0
pydantic>=2.0.0

# Utilities
tenacity>=8.0.0
tqdm>=4.60.0
python-dotenv>=1.0.0
